feat(cli): implement search

+++ daemon/src/controller.rs
@@ -448,11 +448,12 @@ impl MusicPlayer for MusicPlayerServer {
 
     /// returns a list of artists, albums, and songs matching the given search query.
     #[instrument]
+    #[allow(clippy::type_complexity)]
     async fn search(
         self,
         context: Context,
         query: String,
-        limit: i64,
+        limit: u32,
     ) -> (Box<[Song]>, Box<[Album]>, Box<[Artist]>) {
         info!("Searching for: {}", query);
         // basic idea:
@@ -460,19 +461,19 @@ impl MusicPlayer for MusicPlayerServer {
         // 2. search for albums
         // 3. search for artists
         // 4. return the results
-        let songs = Song::search_by_title(&self.db, &query, limit)
+        let songs = Song::search_by_title(&self.db, &query, i64::from(limit))
             .await
             .tap_err(|e| warn!("Error in search: {e}"))
             .unwrap_or_default()
             .into();
 
-        let albums = Album::search_by_title(&self.db, &query, limit)
+        let albums = Album::search_by_title(&self.db, &query, i64::from(limit))
             .await
             .tap_err(|e| warn!("Error in search: {e}"))
             .unwrap_or_default()
             .into();
 
-        let artists = Artist::search_by_name(&self.db, &query, limit)
+        let artists = Artist::search_by_name(&self.db, &query, i64::from(limit))
             .await
             .tap_err(|e| warn!("Error in search: {e}"))
             .unwrap_or_default()
@@ -481,9 +482,9 @@ impl MusicPlayer for MusicPlayerServer {
     }
     /// returns a list of artists matching the given search query.
     #[instrument]
-    async fn search_artist(self, context: Context, query: String, limit: i64) -> Box<[Artist]> {
+    async fn search_artist(self, context: Context, query: String, limit: u32) -> Box<[Artist]> {
         info!("Searching for artist: {}", query);
-        Artist::search_by_name(&self.db, &query, limit)
+        Artist::search_by_name(&self.db, &query, i64::from(limit))
             .await
             .tap_err(|e| {
                 warn!("Error in search_artist: {e}");
@@ -493,9 +494,9 @@ impl MusicPlayer for MusicPlayerServer {
     }
     /// returns a list of albums matching the given search query.
     #[instrument]
-    async fn search_album(self, context: Context, query: String, limit: i64) -> Box<[Album]> {
+    async fn search_album(self, context: Context, query: String, limit: u32) -> Box<[Album]> {
         info!("Searching for album: {}", query);
-        Album::search_by_title(&self.db, &query, limit)
+        Album::search_by_title(&self.db, &query, i64::from(limit))
             .await
             .tap_err(|e| {
                 warn!("Error in search_album: {e}");
@@ -505,9 +506,9 @@ impl MusicPlayer for MusicPlayerServer {
     }
     /// returns a list of songs matching the given search query.
     #[instrument]
-    async fn search_song(self, context: Context, query: String, limit: i64) -> Box<[Song]> {
+    async fn search_song(self, context: Context, query: String, limit: u32) -> Box<[Song]> {
         info!("Searching for song: {}", query);
-        Song::search_by_title(&self.db, &query, limit)
+        Song::search_by_title(&self.db, &query, i64::from(limit))
             .await
             .tap_err(|e| {
                 warn!("Error in search_song: {e}");
